update vram requirements records

+++ models/redpajama.py
@@ -41,7 +41,8 @@ def load_model(
             load_in_8bit=mode_8bit,
             load_in_4bit=mode_4bit,
             device_map="auto",
-            trust_remote_code=True
+            trust_remote_code=True,
+            torch_dtype=torch.float16,
         )#.to(global_vars.device)
 
         if not mode_8bit and not mode_4bit:
